% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{gof.test}
\alias{gof.test}
\title{Goodness-of-fit test}
\usage{
gof.test(
  p_list,
  H0_pdist,
  dissim = protest::Lp.dist,
  alpha = 0.05,
  epsilon,
  verbose = T,
  plot = F,
  dissim_min = 0
)
}
\arguments{
\item{p_list}{List of distribution functions.}

\item{H0_pdist}{Distribution function of the null hypothesis.}

\item{dissim}{Dissimilarity function to be applied, default is the L1
distance.}

\item{alpha}{Significance level, default is 0.05. When given two values, a
three-way test is performed.}

\item{epsilon}{Threshold value.}

\item{verbose}{Should the decision be returned as a message? Default if TRUE.}

\item{plot}{Should the test return a plot of the decision for each
(alpha, epsilon)? Default is FALSE}

\item{dissim_min}{Smallest value the dissimilarity can assume, required only
if plot = TRUE. Default is 0.}
}
\value{
Test results of the goodness-of-fit test.
}
\description{
PROTEST version of the goodness-of-fit test. For a list of
distribution functions and a distribution function/parametric family being
tested, it tests the hypothesis that the true distribution of the data is
described in such function/family.
}
\examples{
p_list <- list(pnorm, function(x) pt(x, df = 5), function(x) pt(x, df = 10),
function(x) pt(x, df = 20), function(x) pt(x, df = 30))

my_dissim <- function(pdist1, pdist2){
protest::Lp.dist(pdist1, pdist2, p = 2)
}

gof.test(p_list, H0_pdist = pnorm, dissim = my_dissim, alpha = 0.05,
epsilon = 0.01, verbose = TRUE, plot = TRUE)

}
