% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adherence.R
\name{lm.dist}
\alias{lm.dist}
\title{Linear model L2 distance}
\usage{
lm.dist(f, X, g = function(x) x, vars = "all", add_intercept = F)
}
\arguments{
\item{f}{Regression function that receives x.}

\item{X}{Covariate matrix.}

\item{g}{Link function. Default is the identity function.}

\item{vars}{Either 'all' or a vector of the columns of X to be added to the
test. Default is 'all'.}

\item{add_intercept}{Should the intercept be added to X? Default is FALSE.}
}
\value{
Distance between function f and the linear model specified by X.
}
\description{
L2 distance between a function f and the linear model is closest
to it on the domain of covariance matrix X. While the distance allows for the
use of any probability measure to be applied to X, we use the empirical
distribution for simplicity.
}
\examples{
f <- function(x) sum(x)
set.seed(42)
n <- 100
X <- cbind(1, runif(n, 1, 2), runif(n, -1, 1))
lm.dist(f, X) #f is linear, so distance is close to 0

}
