% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adherence.R
\name{lm.test}
\alias{lm.test}
\title{Linear model test}
\usage{
lm.test(f_list, X, vars = "all", alpha = 0.05, epsilon, verbose = T, plot = F)
}
\arguments{
\item{f_list}{List of regression functions that receives x}

\item{X}{Covariate matrix}

\item{vars}{Either 'all' or a vector of the columns of X to be added to the
test. Default is 'all'.}

\item{alpha}{Significance level, default is 0.05.}

\item{epsilon}{Threshold value}

\item{verbose}{Should the decision be returned as a message? Default if TRUE}

\item{plot}{Should the test return a plot of the decision for each
(alpha, epsilon)? Default is FALSE}
}
\value{
Test results of the linear model test
}
\description{
PROTEST version of a test that checks the adherence of linear
models to data. The empirical distribution of X is used as the probability
measure for simplicity. Results are a function of the significance level
alpha and the threshold epsilon.
}
\examples{
f_list <- list(f1 = function(x) sum(x), f2 = function(x) 2*sum(x))
set.seed(42)
n <- 100
X <- cbind(1, runif(n, 1, 2), runif(n, -1, 1))
lm.test(f_list, X, alpha = 0.05, epsilon = 0.1, verbose = TRUE, plot = TRUE)

}
