% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adherence.R
\name{lm_test}
\alias{lm_test}
\title{Linear model test}
\usage{
lm_test(
  f_list,
  X,
  epsilon,
  g = function(x) x,
  vars = "all",
  add_intercept = FALSE,
  alpha = 0.05,
  verbose = TRUE,
  plot = FALSE
)
}
\arguments{
\item{f_list}{List of regression functions that receives x}

\item{X}{Covariate matrix}

\item{epsilon}{Threshold value.}

\item{g}{Link function. Default is the identity function.}

\item{vars}{Either 'all' or a vector of the columns of X to be added to the
test. Default is 'all'.}

\item{add_intercept}{Should the intercept be added to X? Default is FALSE.}

\item{alpha}{Significance level, default is 0.05. When given two values, a
three-way test is performed.}

\item{verbose}{Should the decision be returned as a message? Default if TRUE.}

\item{plot}{Should the test return a plot of the decision for each
(alpha, epsilon)? Default is FALSE}
}
\value{
Test results of the linear model test
}
\description{
PROTEST version of a test that checks the adherence of linear
models to data. The empirical distribution of X is used as the probability
measure for simplicity. Results are a function of the significance level
alpha and the threshold epsilon.
}
\examples{
f_list <- list(f1 = function(x) sum(x), f2 = function(x) 2 * sum(x))
set.seed(42)
n <- 100
X <- cbind(1, runif(n, 1, 2), runif(n, -1, 1))
lm_test(f_list, X, alpha = 0.05, epsilon = 0.1, verbose = TRUE, plot = TRUE)

}
