% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissim.R
\name{Lp.dist}
\alias{Lp.dist}
\title{Lp distance on the Lebesgue measure.}
\usage{
Lp.dist(pdist1, pdist2, p = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{pdist1}{First distribution function.}

\item{pdist2}{Second distribution function.}

\item{p}{Index of the Lp distance to be used, default is 1 (L1 distance).}

\item{lower}{Lower bound of the variable, default is -Inf}

\item{upper}{Upper bound of the variable, default is -Inf}
}
\value{
Lp distance between two functions.
}
\description{
Lp distance to be used for comparing distribution functions.
}
\examples{
Lp.dist(pdist1 = pnorm, pdist2 = function(x) pt(x, df = 10), p=2)

}
