% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quant.dissim}
\alias{quant.dissim}
\title{L1 distance for the quantile test}
\usage{
quant.dissim(pdist, qdist, x0 = 0, p0 = 0.5)
}
\arguments{
\item{pdist}{Distribution function that receives x.}

\item{qdist}{Quantile distribution associated to pdist.}

\item{x0}{Value from the null hypothesis, default is 0.}

\item{p0}{Probability from the null hypothesis, default is 0.5.}
}
\value{
Distance between distribution function F and the class of
distribution functions such that G(x0) = p0.
}
\description{
L1 distance between a distribution function F and the function G
that best appoximates and obeys the condition G(x0) = p0.
}
\examples{
quant.dissim(pdist = pexp, qdist = qexp, x0 = 1, p0 = 0.5)

}
