% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{test.results}
\alias{test.results}
\title{PROTEST results}
\usage{
test.results(
  dissim,
  alpha = 0.05,
  epsilon,
  verbose = T,
  plot = F,
  dissim_min = 0
)
}
\arguments{
\item{dissim}{Vector of dissimilarities}

\item{alpha}{Significance level, default is 0.05}

\item{epsilon}{Threshold}

\item{verbose}{Should the decision be returned as a message? Default if TRUE}

\item{plot}{Should the test return a plot of the decision for each
(alpha, epsilon)? Default is FALSE}

\item{dissim_min}{Smallest value possible for the dissimilarity function,
default is 0.}
}
\value{
Test results (decision, dissimilarity vector, significance level,
threshold and plot).
}
\description{
Test results based on the dissimilarity, significance alpha and
threshold epsilon. The procedure can also return a plot that shows the
decision based on each combination of (alpha, epsilon).
}
\examples{
dissim <- seq(.5, 1, length.out = 100)
test.results(dissim = dissim, epsilon = .5, verbose = TRUE, plot = TRUE, dissim_min = 0.5)

}
