% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{test_results}
\alias{test_results}
\title{PROTEST results}
\usage{
test_results(
  dissim,
  epsilon,
  alpha = 0.05,
  verbose = TRUE,
  plot = FALSE,
  dissim_min = 0,
  three_way = FALSE
)
}
\arguments{
\item{dissim}{Vector of dissimilarities.}

\item{epsilon}{Threshold.}

\item{alpha}{Significance level, default is 0.05.}

\item{verbose}{Should the decision be returned as a message? Default is TRUE.}

\item{plot}{Should the test return a plot of the decision for each
(alpha, epsilon)? Default is FALSE.}

\item{dissim_min}{Smallest value possible for the dissimilarity function,
default is 0.}

\item{three_way}{Should the results provided be of a three-way test? Default
is FALSE. If two values for alpha are provided, this option is ignored.}
}
\value{
Test results (decision, dissimilarity vector, significance level,
threshold and plot). If plotting on the three-way case, the plot assumes that
alpha2 = 1 - alpha1.
}
\description{
Test results based on the dissimilarity, significance alpha and
threshold epsilon. The procedure can also return a plot that shows the
decision based on each combination of (alpha, epsilon).
}
\examples{
dissim <- seq(.5, 1, length.out = 100)
test_results(
  dissim = dissim,
  epsilon = .8,
  verbose = TRUE, plot = TRUE, dissim_min = 0.5
)

}
