% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample.R
\name{twosample.test}
\alias{twosample.test}
\title{Two-sample test}
\usage{
twosample.test(
  p1_list,
  p2_list,
  dist = function(pdist1, pdist2) {     protest::Lp.dist(pdist1, pdist2, p = 1) },
  alpha = 0.05,
  epsilon,
  verbose = T,
  plot = F
)
}
\arguments{
\item{p1_list}{List of distribution functions based on the first variable.}

\item{p2_list}{List of distribution functions based on the first variable.}

\item{dist}{Distance function to be applied, default is the L1 distance.}

\item{alpha}{Significance level, default is 0.05. When given two values, a
three-way test is performed.}

\item{epsilon}{Threshold value.}

\item{verbose}{Should the decision be returned as a message? Default if TRUE.}

\item{plot}{Should the test return a plot of the decision for each
(alpha, epsilon)? Default is FALSE}
}
\value{
Test results of the two-sample test.
}
\description{
PROTEST version of the two-sample test. For two list of
distribution functions, one for each datasets, it tests the hypothesis that
the true distribution function is the same for both sets.
}
\examples{
p1_list <- list(pexp,
function(x) pgamma(x, shape = 2),
function(x) pexp(x, rate = 2))

p2_list <- list(function(x) pexp(x, rate = 1.1),
function(p) pgamma(p, shape = 2.1),
function(p) pexp(p, rate = 2.1))

my_dist <- function(pdist1, pdist2) protest::Lp.dist(pdist1, pdist2, p = 2)

twosample.test(p1_list, p2_list, dist = my_dist, alpha = 0.05, epsilon = 0.1,
verbose = TRUE, plot = TRUE)

}
